requirements.txt
openai
python-dotenv
rich
numpy
sqlite3
tqdm
pickle
termcolor
PyPDF2
python-docx